+++ core/07-unpack.php
@@ -7,9 +7,11 @@
 // If the extra data directory:
 //  - doesn't exist already
 //  - has an mtime before that of this file
+//  - is empty
 // ...extract it again
 if(!file_exists($paths->extra_data_directory) || 
-	filemtime(__FILE__) > filemtime($paths->extra_data_directory)) {
+	filemtime(__FILE__) > filemtime($paths->extra_data_directory)
+	|| is_directory_empty($paths->extra_data_directory)) {
 	
 	$error_message_help = "<p>Have you checked that PHP has write access to the directory that <code>index.php</code> is located in (and all it's contents and subdirectories)? Try <code>sudo chown USERNAME:USERNAME -R path/to/directory</code> and <code>sudo chmod -R 0644 path/to/directory; sudo chmod -R +X path/too/directory</code>, where <code>USERNAME</code> is the username that the PHP process is running under.</p>";
 	
@@ -44,6 +46,11 @@ if(!file_exists($paths->extra_data_directory) ||
 	fclose($temp_file);
 	
 	$extractor = new ZipArchive();
+	if(!class_exists("ZipArchive") || !($extractor instanceof ZipArchive)) {
+		if(file_exists($paths->extra_data_directory))
+			delete_recursive($paths->extra_data_directory);
+		exit(page_renderer::render_minimal("Unpacking error - $settings->sitename", "<p>Oops! $settings->sitename wasn't able to unpack itself because the ZipArchive doesn't exist or is faulty. Please install the PHP zip extension (on apt-based systems it's the <code>php-zip</code> package) and then try again later. You can check that it's installed by inspecting the output of <code>php -m</code>, or running the <a href='https://www.php.net/manual/en/function.phpinfo.php'><code>phpinfo()</code> command</a>."));
+	}
 	$extractor->open($temp_filename);
 	$extractor->extractTo($paths->extra_data_directory);
 	$extractor->close();